refactor: move reporter_lifecycle into pixi_compute_reporters
