+++ crates/pixi_command_dispatcher/src/instantiate_backend_key.rs
@@ -41,10 +41,9 @@ use crate::discovered_backend::DiscoveredBackendKey;
 use crate::ephemeral_env::{EphemeralEnvError, EphemeralEnvKey, EphemeralEnvSpec};
 use crate::injected_config::ToolBuildEnvironmentKey;
 use crate::reporter::InstantiateBackendReporter;
-use crate::reporter_lifecycle::{Active, LifecycleKind, ReporterLifecycle};
 use crate::resolved_backend_command::{ResolvedBackendCommand, ResolvedBackendCommandKey};
 use pixi_compute_cache_dirs::CacheDirsKey;
-use pixi_compute_reporters::OperationId;
+use pixi_compute_reporters::{Active, LifecycleKind, OperationId, ReporterLifecycle};
 
 /// CLI-driven overrides applied to the discovered project model before
 /// the backend is initialized. Each field is content-addressed into
